re-enabled loginstate
For some reason, if the client-state instance variable in client-class is type-constrained to require a value of type FabricClientState, then LoginState values cannot be stored in it. Other subclasses of FabricClientState can, but not LoginState. I have therefore temporarily removed the type constraint, which of course causes the compiler to complain about undefined methods